Refactors the Factory structure (modularized)
