+++ tests/e2e/factory/resources/index.js
@@ -1,0 +1,7 @@
+import { createArticle } from './articles'
+import { createMagazine } from './magazines'
+
+export {
+  createArticle,
+  createMagazine,
+}
