+++ tests/e2e/factory/store/index.js
@@ -1,7 +1,7 @@
 import createInitialVuexStoreState from './initial.state'
 import createInitialVuexStoreGetters from './initial.getters'
 
-export const store = {
+export {
   createInitialVuexStoreState,
   createInitialVuexStoreGetters
 }
